[add] added cs files to devide class
