+++ Assets/GameLogic/DataTemplates.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 9aca7b40c80568f4384d711f1c670363
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
