+++ Assets/GameLogic/DataTemplates/NPCDataTemplate.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 8309e93f7b1b4964f8c8086ed1be660a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
