+++ Assets/GameLogic/DataTemplates/SkillsDataTemplate.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 714c2b12b18395547a5df9c884a598d7
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
