+++ Assets/Shop/Shop.cs
@@ -24,12 +24,17 @@ namespace TeamC
             // ★NPCをシーン場へInstanciateする処理はSuperクラスには書いていない★
         }
 
+        void TaskToUpdateText(decimal cost, int boughtCount)
+        {
+            
+        }
+
         /// <summary> ボタンからNPC名を渡して購入時にこれをボタンから呼び出す </summary>
         public void BuyNPC(string name)
         {
             // Superクラスでは、購入数に応じてコストを算出し、
             // それをプレイヤーへコストの適応をして、購入数を＋１しただけ
-            base.DecreasePlayerSource(name, CalculateNPCCost(name), TaskToInstantiateNPC);
+            base.DecreasePlayerSource(name, CalculateNPCCost(name), TaskToInstantiateNPC, TaskToUpdateText);
         }
 
         /// <summary> NPCの購入数に応じた価格の算出 </summary>
